+++ docente/pianoDiLavoroDuplica.php
@@ -15,8 +15,8 @@ if(isset($_POST)) {
 	$original_piano_di_lavoro_id = $_POST['id'];
 
     // legge il contenuto del record corrente e lo duplica
-    $query = "INSERT INTO piano_di_lavoro(docente_id, materia_id, anno_scolastico_id, indirizzo_id, classe, sezione, nome_classe, template, stato, competenze, note_aggiuntive, carenza)
-            SELECT '$__docente_id', materia_id, '$__anno_scolastico_corrente_id', indirizzo_id, classe, sezione, nome_classe, 0, 'draft', competenze, note_aggiuntive, carenza
+    $query = "INSERT INTO piano_di_lavoro(docente_id, materia_id, anno_scolastico_id, indirizzo_id, classe, sezione, nome_classe, template, obiettivi_minimi, stato, competenze, note_aggiuntive, carenza)
+            SELECT '$__docente_id', materia_id, '$__anno_scolastico_corrente_id', indirizzo_id, classe, sezione, nome_classe, 0, 0, 'draft', competenze, note_aggiuntive, carenza
             FROM piano_di_lavoro WHERE id = $original_piano_di_lavoro_id; ";
     dbExec($query);