+++ docente/pianoDiLavoroSave.php
@@ -55,6 +55,7 @@ if(isset($_POST)) {
 	$classe = $_POST['classe'];
 	$sezione = $_POST['sezione'];
 	$template = $_POST['template'];
+	$obiettivi_minimi = $_POST['obiettivi_minimi'];
 	$clil = $_POST['clil'];
 	$stato = $_POST['stato'];
     $competenze = escapePost('competenze');
@@ -77,7 +78,7 @@ if(isset($_POST)) {
     }
     
     if ($id > 0) {
-        $query = "UPDATE piano_di_lavoro SET docente_id = $docente_id, materia_id = $materia_id, anno_scolastico_id = $anno_scolastico_id, indirizzo_id = $indirizzo_id, classe = $classe, sezione = '$sezione', nome_classe = '$nome_classe', template = $template, clil = $clil, stato = '$stato', competenze = '$competenze' , note_aggiuntive = '$note_aggiuntive' ".($studente_id != null? ", studente_id = $studente_id" : "")." WHERE id = '$id' ;";
+        $query = "UPDATE piano_di_lavoro SET docente_id = $docente_id, materia_id = $materia_id, anno_scolastico_id = $anno_scolastico_id, indirizzo_id = $indirizzo_id, classe = $classe, sezione = '$sezione', nome_classe = '$nome_classe', template = $template, obiettivi_minimi = $obiettivi_minimi, clil = $clil, stato = '$stato', competenze = '$competenze' , note_aggiuntive = '$note_aggiuntive' ".($studente_id != null? ", studente_id = $studente_id" : "")." WHERE id = '$id' ;";
         dbExec($query);
         info("aggiornato piano_di_lavoro id=$id");
 
@@ -105,7 +106,7 @@ if(isset($_POST)) {
             salvaTic($id, $tic_id_list_str);
         }
     } else {
-        $query = "INSERT INTO piano_di_lavoro(docente_id, materia_id, anno_scolastico_id, indirizzo_id, classe, sezione, nome_classe, template, clil, stato, competenze, note_aggiuntive, carenza".($studente_id != null? ", studente_id" : "").") VALUES($docente_id, $materia_id, $anno_scolastico_id, $indirizzo_id, '$classe', '$sezione', '$nome_classe', $template, $clil, '$stato', '$competenze', '$note_aggiuntive', $carenza".($studente_id != null? ", $studente_id" : "").")";
+        $query = "INSERT INTO piano_di_lavoro(docente_id, materia_id, anno_scolastico_id, indirizzo_id, classe, sezione, nome_classe, template, obiettivi_minimi, clil, stato, competenze, note_aggiuntive, carenza".($studente_id != null? ", studente_id" : "").") VALUES($docente_id, $materia_id, $anno_scolastico_id, $indirizzo_id, '$classe', '$sezione', '$nome_classe', $template, $obiettivi_minimi, $clil, '$stato', '$competenze', '$note_aggiuntive', $carenza".($studente_id != null? ", $studente_id" : "").")";
         dbExec($query);
         $id = dblastId();
         info("aggiunto piano_di_lavoro id=$id");